+++ commit-scoreboard/features/steps/scoring.py
@@ -1,51 +1,16 @@
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from typing import Any, NamedTuple
 from unittest.mock import Mock
 from behave import given, when, then
 
-class VersionControlRepo(ABC):
-    @abstractmethod
-    def count_commits_ahead(self, ref: str, target: str) -> int: ...
-
-class DummyVersionControlRepo(VersionControlRepo):
-    def __init__(self, **commits: dict[str, int]) -> None:
-        self.branch_commits = commits
-
-    def count_commits_ahead(self, ref: str, target: str) -> int:
-        return self.branch_commits[target] - self.branch_commits[ref]
-        
-
-class Score(NamedTuple):
-    score: int
-
-class ScoreboardPresenter(ABC):
-    @abstractmethod
-    def present(self, scores: dict[str, Any]) -> None: ...
-
-
-@dataclass(frozen=True)
-class CalculateScores:
-    version_control_repo: VersionControlRepo
-    scoreboard: ScoreboardPresenter
-
-    def run(self, ref_branch: str, target_branches: list[str]) -> None:
-        target_branch = target_branches[0]
-        commits_ahead = self.version_control_repo.count_commits_ahead(ref=ref_branch, target=target_branch)
-        scores = {target_branch: Score(score=commits_ahead)}
-        self.scoreboard.present(scores=scores)
-
-        
+from scoreboard.workflow import CalculateScores, ScoreboardPresenter
+from scoreboard.vcs_repos import DummyVersionControlRepo
     
-
-
+     
 @given(u'the {team} branch is {n:d} commits ahead of the reference {ref} branch')
 def step_impl(context, team, ref, n):
     branches = {ref: 0, team: n}
     context.vcs = DummyVersionControlRepo(**branches)
     
     
-
 @when(u'the scores are calculated for teams {team} against reference branch {ref}')
 def step_impl(context, team, ref):
     context.scoreboard = Mock(ScoreboardPresenter)
